requirements.txt
streamlit
Pillow
requests
python-dotenv
